Fixed some MS-Edge issues (#90)
Replaced use of Array.flat
"Fixed" `withoutOpacity` function so that it uses "rgb()" instead of "rgba()"
Added workaround for weird Edge issue that doesn't render the plot-area of some charts (bar/column charts and heatmap) unless you resize the window. An immediate redraw of the D3 element seems to work, but is not great.