Improved and simplified nearby-tooltip code and used on y-scatter
It turns out we don't actually need to use copies of the scales, and therefore can avoid decorating the chart object to get the "measure" (though this approach does require calling the "toolTip" component after the chart has been rendered. This allows us to simplify the code quite a lot.
Also:
* rebind `settings` and other properties from base tooltip component
* added `size` option for controlling highlight sizing (though the "range" has to be modified because the symbol size behaviour isn't the same as the circle radius behaviour)
* Use nearby-tooltip for the y-scatter chart, since it's a better experience