Remove webpack as a peerDependency for the webpack plugin
This PR removes webpack as a peerDependency for `@finos/perspective-webpack-plugin`

A peer dependency (or dependency) should only be included if it's actually required by the source code. (For the same reason, `file-loader` and `worker-loader` _should_ be included because they're implicitly required.) Keeping this peerDependency makes it so that the consuming package has to declare webpack as a dependency, even if they're only passing the plugin to a tool that transitively uses webpack.