Add macos-14 to test_python matrix
macos-14 is aarch64 with Apple silicon.  This lets us test the aarch64
wheel in CI

Will try a follow up to see if the builds are any faster on macos-14.
Will want to double check we still have a good value for
`CMAKE_OSX_DEPLOYMENT_TARGET`/`MACOSX_DEPLOYMENT_TARGET` after upgrading
the builders

Signed-off-by: Tom Jakubowski <tom@prospective.dev>

new crack at build matrix

verified this one with a model of the behavior in github's docs.  would
be cool if I knew how to get github to print the job configurations
before the job actually runs